pyproject.toml
[tool.poetry]
name = "genesis-ai"
version = "1.0.0"
description = "GENESIS - The Next Evolution in AI Systems"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "genesis"}]

[tool.poetry.dependencies]
python = "^3.9"
tensorflow = "^2.8.0"
numpy = "^1.21.0"
pandas = "^1.4.0"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
psutil = "^5.9.0"
pyyaml = "^6.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^22.0.0"
flake8 = "^4.0.0"
mypy = "^1.0.0"
pytest-cov = "^4.1.0"
bandit = "^1.7.0"
isort = "^5.12.0"
pre-commit = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=genesis --cov-report=term-missing" 